Feat: Components changet to be more flexible. Can be used on diferent pages
