+++ src/components/addExerciseTemplate/AddExerciseTemplate.css
@@ -1,0 +1,10 @@
+.addExercisePanel-wrapper {
+  margin-top: 20px;
+  width: 100%;
+  height: 20%;
+  box-sizing: border-box;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-wrap: wrap;
+}
