--- src/components/createTrainingPanel/AddExercisePanel.css
@@ -1,30 +1,0 @@
-.addExercisePanel-wrapper {
-  margin-top: 20px;
-  width: 100%;
-  height: 20%;
-  box-sizing: border-box;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-wrap: wrap;
-}
-
-.addExercisePanel-input-wrapper {
-  position: relative;
-  width: fit-content;
-  height: fit-content;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 10px;
-}
-
-.addExercisePanel-input {
-  height: 20px;
-  margin: 0;
-  padding: 0;
-}
-
-.addExercisePanel-btn {
-  height: 20px;
-}
