+++ src/components/exampleExercisesList/ExampleExercisesList.css
@@ -1,3 +1,23 @@
+.exampleExercise-input-wrapper {
+  position: relative;
+  width: fit-content;
+  height: fit-content;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  gap: 10px;
+}
+
+.exampleExercise-input {
+  height: 20px;
+  margin: 0;
+  padding: 0;
+}
+
+.exampleExercise-btn {
+  height: 20px;
+}
+
 .exampleExercisesList-wrapper {
   display: flex;
   position: absolute;
