+++ src/components/exerciseTemplatePanel/ExerciseTemplatePanel.css
