+++ src/components/exerciseTemplatePanel/ExerciseTemplatePanel.jsx
