+++ src/components/weekPlanPanel/WeekPlanPanel.css
@@ -1,0 +1,16 @@
+.weekPlanPanel-wrapper {
+  border: 3px solid rgb(180, 179, 179);
+  border-radius: 30px;
+  min-width: 40%;
+  width: fit-content;
+  height: 600px;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  box-shadow: -20px 20px 40px rgba(0, 0, 0, 0.25);
+  background-color: rgb(247, 246, 246);
+}
+
+.weekPlan-addTraining-btn {
+  margin-top: 20px;
+}
