+++ src/pages/WeekPlanPage.jsx
@@ -1,5 +1,6 @@
 import React from "react";
+import { WeekPlanPanel } from "../components/weekPlanPanel/WeekPlanPanel";
 
 export const WeekPlanPage = () => {
-  return <p>WeekPlanPage</p>;
+  return <WeekPlanPanel />;
 };
